refactor(release): give the secret scanner its own module
The export ratchet was right: a same-file caller is not a cross-file consumer, and the scanner is a distinct concern from the validator that reports it.
